refactor(logic): split simple AI heuristics test coverage into smaller files
Keep the existing simple AI heuristic assertions intact while splitting generation tests into a separate file to reduce test file size and continue the issue #2288 size-debt cleanup in reviewable slices.

Refs #2288
